feat(sda-commons-shared-tracing): deprecate tracing module in favor of same API provided by sda-commons-server-dropwizard
