+++ sda-commons-shared-tracing/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
+  implementation project(':sda-commons-server-dropwizard')
   implementation 'org.slf4j:slf4j-api'
   testImplementation project(':sda-commons-server-testing')
 }
